NewBoosteroidCode/requirements.txt
asyncpraw~=7.8.1
aiohttp>=3.6.0,<3.8.0
certifi~=2024.12.14
asyncio~=3.4.3
notion-client~=2.3.0
python-dotenv~=1.0.1
python-telegram-bot~=21.10
discord.py~=1.7.3
nest-asyncio~=1.6.0